update: courseSeq int되어있던거->long 변경,장바구니 담기 완료, redirect 후 모달알림

+++ src/main/java/com/up/spring/coach/controller/CoachController.java
@@ -109,7 +109,7 @@ public class CoachController {
     }
 
     @RequestMapping("/insertCurriculum")
-    public String insertCurriculum(Curriculum curriculum, Model model,int courseSeq, HttpSession session) {
+    public String insertCurriculum(Curriculum curriculum, Model model,long courseSeq, HttpSession session) {
         if ("UPLOAD".equals(curriculum.getCurrVideoType()) && curriculum.getCurrVideoFile() != null) {
             MultipartFile file = curriculum.getCurrVideoFile();
             if (!file.isEmpty()) {
